+++ tests/test_archive_files.py
@@ -1,12 +1,9 @@
-"""Tests for epicsarchiver.utils module."""
 import logging
 import os
-from datetime import datetime
 
 import pytest
-from pytz import UTC
 
-from epicsarchiver import utils
+from epicsarchiver import archive_files
 
 SAMPLES_PATH = os.path.join(os.path.abspath(os.path.dirname(__file__)), "samples")
 FILE1_PVS = [
@@ -32,31 +29,22 @@ FILE2_RENAME = [
 ]
 
 
-def test_format_date():
-    assert utils.format_date("20180715") == "2018-07-15T00:00:00.000000Z"
-    assert utils.format_date("20180715 17:45") == "2018-07-15T17:45:00.000000Z"
-    assert (
-        utils.format_date(datetime(2018, 7, 15, 19, 5, tzinfo=UTC))
-        == "2018-07-15T19:05:00.000000Z"
-    )
-
-
-def test_parse_archive_file():
+def test_parse_archive_file() -> None:
     filename = os.path.join(SAMPLES_PATH, "file1.archive")
-    pvs = utils.parse_archive_file(filename)
+    pvs = archive_files.parse_archive_file(filename)
     assert list(pvs) == FILE1_PVS
 
 
-def test_parse_rename_file():
+def test_parse_rename_file() -> None:
     filename = os.path.join(SAMPLES_PATH, "file1.rename")
-    pvs = utils.parse_rename_file(filename)
+    pvs = archive_files.parse_rename_file(filename)
     assert list(pvs) == FILE1_RENAME
 
 
-def test_parse_rename_file_incomplete_line(caplog):
+def test_parse_rename_file_incomplete_line(caplog: pytest.LogCaptureFixture) -> None:
     filename = os.path.join(SAMPLES_PATH, "file2.rename")
     with caplog.at_level(logging.ERROR):
-        pvs = utils.parse_rename_file(filename)
+        pvs = archive_files.parse_rename_file(filename)
     assert list(pvs) == FILE2_RENAME
     captured_log = caplog.text
     assert (
@@ -65,49 +53,25 @@ def test_parse_rename_file_incomplete_line(caplog):
     )
 
 
-def test_get_pvs_from_files():
+def test_get_pvs_from_files() -> None:
     files = [
         os.path.join(SAMPLES_PATH, "file1.archive"),
         os.path.join(SAMPLES_PATH, "file2.archive"),
     ]
-    pvs = utils.get_pvs_from_files(files)
+    pvs = archive_files.get_pvs_from_files(files)
     assert pvs == FILE1_PVS + FILE2_PVS
 
 
-def test_get_pvs_from_files_with_appliance():
+def test_get_pvs_from_files_with_appliance() -> None:
     files = [os.path.join(SAMPLES_PATH, "file2.archive")]
-    pvs = utils.get_pvs_from_files(files, appliance="appliance0")
+    pvs = archive_files.get_pvs_from_files(files, appliance="appliance0")
     assert pvs == FILE2_PVS_APPLIANCE
 
 
-def test_get_rename_pvs_from_files():
+def test_get_rename_pvs_from_files() -> None:
     files = [
         os.path.join(SAMPLES_PATH, "file1.rename"),
         os.path.join(SAMPLES_PATH, "file2.rename"),
     ]
-    pvs = utils.get_rename_pvs_from_files(files)
+    pvs = archive_files.get_rename_pvs_from_files(files)
     assert pvs == FILE1_RENAME + FILE2_RENAME
-
-
-@pytest.mark.parametrize(
-    "test_input,expected",
-    [({"status": "ok"}, True), ({"status": "foo"}, False), ({"hello": "world"}, False)],
-)
-def test_check_result(test_input, expected):
-    output = utils.check_result(test_input)
-    assert output is expected
-
-
-@pytest.mark.parametrize(
-    "test_input,default_message,output",
-    [
-        ({"status": "nok"}, "Not OK", "Not OK\n"),
-        ({"validation": "Hello"}, None, "Hello\n"),
-        ({"validation": "Hello"}, "foo", "Hello\n"),
-    ],
-)
-def test_check_result_message(caplog, test_input, default_message, output):
-    with caplog.at_level(logging.ERROR):
-        utils.check_result(test_input, default_message)
-    captured_log = caplog.text
-    assert output in captured_log