+++ tests/test_command.py
@@ -1,10 +1,14 @@
 """Tests for epicsarchiver.command module."""
+from pathlib import Path
+
 from click.testing import CliRunner
+from pytest import CaptureFixture
+from pytest_mock import MockerFixture
 
 from epicsarchiver import command
 
 
-def test_archive_file_does_not_exist():
+def test_archive_file_does_not_exist() -> None:
     runner = CliRunner()
     files = ["file1", "file2"]
     result = runner.invoke(command.cli, ["archive", *files])
@@ -12,12 +16,14 @@ def test_archive_file_does_not_exist():
     assert "Path 'file1' does not exist." in result.output
 
 
-def test_archive_file_does_exist(tmpdir, mocker, capsys):
+def test_archive_file_does_exist(
+    tmp_path: Path, mocker: MockerFixture, capsys: CaptureFixture
+) -> None:
     mock_archiver = mocker.patch("epicsarchiver.command.ArchiverAppliance")
-    file1 = tmpdir.join("file1")
-    file1.write("test")
-    file2 = tmpdir.join("file2")
-    file2.write("test")
+    file1 = tmp_path.joinpath("file1")
+    file1.open("w").write("test")
+    file2 = tmp_path.joinpath("file2")
+    file2.open("w").write("test")
     runner = CliRunner()
     files = (str(file1), str(file2))
     with capsys.disabled():
@@ -30,11 +36,13 @@ def test_archive_file_does_exist(tmpdir, mocker, capsys):
         mock_archiver.assert_called_once_with("localhost")
 
 
-def test_archive_hostname(tmpdir, mocker, capsys):
+def test_archive_hostname(
+    tmp_path: Path, mocker: MockerFixture, capsys: CaptureFixture
+) -> None:
     mock_archiver = mocker.patch("epicsarchiver.command.ArchiverAppliance")
     hostname = "myarchiver.example.org"
-    file1 = tmpdir.join("file1")
-    file1.write("test")
+    file1 = tmp_path.joinpath("file1")
+    file1.open("w").write("test")
     runner = CliRunner()
     with capsys.disabled():
         result = runner.invoke(
@@ -44,11 +52,13 @@ def test_archive_hostname(tmpdir, mocker, capsys):
         mock_archiver.assert_called_once_with(hostname)
 
 
-def test_archive_with_appliance(tmpdir, mocker, capsys):
+def test_archive_with_appliance(
+    tmp_path: Path, mocker: MockerFixture, capsys: CaptureFixture
+) -> None:
     mock_archiver = mocker.patch("epicsarchiver.command.ArchiverAppliance")
     appliance = "foo"
-    file1 = tmpdir.join("file1")
-    file1.write("test")
+    file1 = tmp_path.joinpath("file1")
+    file1.open("w").write("test")
     runner = CliRunner()
     with capsys.disabled():
         result = runner.invoke(
@@ -61,13 +71,13 @@ def test_archive_with_appliance(tmpdir, mocker, capsys):
         )
 
 
-def test_rename(tmpdir, mocker):
+def test_rename(tmp_path: Path, mocker: MockerFixture) -> None:
     mock_archiver = mocker.patch("epicsarchiver.command.ArchiverAppliance")
     hostname = "myarchiver.example.org"
-    file1 = tmpdir.join("file1")
-    file1.write("test")
-    file2 = tmpdir.join("file2")
-    file2.write("test2")
+    file1 = tmp_path.joinpath("file1")
+    file1.open("w").write("test")
+    file2 = tmp_path.joinpath("file2")
+    file2.open("w").write("test2")
     runner = CliRunner()
     result = runner.invoke(
         command.cli, ["--hostname", hostname, "rename", str(file1), str(file2)]